backend/requirements.txt
django>=5.0
psycopg[binary]
djangorestframework
pytest
django-parler
pytest-django
